collection created

+++ backend/src/middlewares/validator.middlewares.js
@@ -1,43 +1,44 @@
-const { deleteFile } = require("../utilies/helper")
+const { deleteFile } = require("../utilies/helper");
 
+const bodyValidator = (schema) => {
+    return async (req, res, next) => {
+        try {
+            const data = req.body;
+            if (req.file) {
+                data[req.file.fieldname] = req.file.filename;
+            }
 
-const bodyValidator = (schema)=>{
-    return async (req,res,next)=>{
-       try {
-        const data = req.body
-        if(req.file){
-            data[req.file.fieldname] = req.file.filename
+            await schema.validateAsync(data, { abortEarly: false });
+            next();
+        } catch (exception) {
+            const detail = {};
+
+            // Check if Joi validation error
+            if (exception.isJoi && exception.details) {
+                // File delete
+                if (req.file) {
+                    console.log(req.file);
+                    deleteFile("./" + req.file.path);
+                }
+
+                // Map the validation errors to details object
+                exception.details.map((error) => {
+                    console.log(error);
+                    detail[error["path"][0]] = error.message;
+                });
+            } else {
+                // Handle other types of errors (e.g., unexpected errors)
+                console.error(exception);
+            }
+
+            next({
+                status: 400,
+                details: detail.length > 0 ? detail : { error: 'An unknown error occurred' }
+            });
         }
-
-        
-    
-        await schema.validateAsync(data, {abortEarly: false})
-        next();
-    }
-    catch(exception){
-           const detail = {};
-
-           //file delete
-           if(req.file){
-            console.log(req.file) 
-            deleteFile("./"+req.file.path)
-           }
-
-           exception.details.map((error) =>{
-            console.log(error) 
-            detail[error["path"][0]] = error.message
-           })
-       next({
-        status:400,
-        details:detail
-       }) 
-    }
-    }
-    
-
-    
-}
+    };
+};
 
 module.exports = {
     bodyValidator
-}
+};
